restructure pages under /app
