+++ app/users/users-table.tsx
@@ -1,6 +1,6 @@
 import { timeAgo } from '@/lib/utils';
 import Image from 'next/image';
-import { RefreshButton } from '../buttons';
+import { RefreshButton } from '@/components/buttons';
 import * as rpc from '../../rpc';
 
 /**
@@ -8,25 +8,22 @@ import * as rpc from '../../rpc';
  */
 export async function UsersTable() {
   let data;
-  const startTime = Date.now();
-
   try {
     data = await rpc.users.getAll();
   } catch (e) {
     throw new Error(`get users failed:` + e);
   }
-
   const { rows: users } = data;
-  const duration = Date.now() - startTime;
 
+  if (!users) {
+    return <>Loading...</>;
+  }
   return (
     <div className="bg-white/30 p-12 shadow-xl ring-1 ring-gray-900/5 rounded-lg backdrop-blur-lg max-w-xl mx-auto w-full">
       <div className="flex justify-between items-center mb-4">
         <div className="space-y-1">
           <h2 className="text-xl font-semibold">Recent Users</h2>
-          <p className="text-sm text-gray-500">
-            Fetched {users.length} users in {duration}ms
-          </p>
+          <p className="text-sm text-gray-500">Fetched {users.length} users.</p>
         </div>
         <RefreshButton />
       </div>
@@ -47,6 +44,34 @@ export async function UsersTable() {
               </div>
             </div>
             <p className="text-sm text-gray-500">{timeAgo(user.createdAt)}</p>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}
+
+export function UsersTablePlaceholder() {
+  return (
+    <div className="bg-white/30 p-12 shadow-xl ring-1 ring-gray-900/5 rounded-lg backdrop-blur-lg max-w-xl mx-auto w-full">
+      <div className="flex justify-between items-center mb-4">
+        <div className="space-y-1">
+          <h2 className="text-xl font-semibold">Recent Users</h2>
+          <p className="text-sm text-gray-500">Fetching users...</p>
+        </div>
+        <RefreshButton />
+      </div>
+      <div className="divide-y divide-gray-900/5">
+        {[...Array(3)].map((_, i) => (
+          <div key={i} className="flex items-center justify-between py-3">
+            <div className="flex items-center space-x-4">
+              <div className="h-12 w-12 rounded-full bg-gray-200 animate-pulse" />
+              <div className="space-y-1">
+                <div className="h-6 w-28 rounded-md bg-gray-200 animate-pulse" />
+                <div className="h-4 w-24 rounded-md bg-gray-200 animate-pulse" />
+              </div>
+            </div>
+            <div className="h-4 w-12 rounded-md bg-gray-200 animate-pulse" />
           </div>
         ))}
       </div>
