--- components/users/index.ts
@@ -1,2 +1,0 @@
-export * from './users-table-placeholder';
-export * from './users-table';
